file_contexts: Label /dev/ttyS2 as console_device
This fixes the issue with the emulator "-shell" option.
Init tries to open the console which is passed through
the kernel androidboot.console property, but fails to
open it because "avc" denies it. Init only has permissions
to open console_device in rw mode. This ensures that
/dev/ttyS2 is properly labeled as console_device.

Replaced tabs with spaces.

(cherry picked from commit df2620ada4155e4123952500cf148f9b2aefe3c1)

Change-Id: I4b5994ee8539a8394fe697145982d48e401160ff