MIPS: Use current NDK version for mips32r6
This patch reverts the workaround earlier put to switch to the NDK R10
when compiling for mips32r6, since only that version contained all the
libraries generated for mips32r6, until now.

Now that we have full mips32r6 support in the currently default NDK r13,
switch to using those.

Change-Id: I4c3d0a6d7e3780b9bd21b02d9118784630ee1337
Signed-off-by: Miodrag Dinic <miodrag.dinic@imgtec.com>